solve: overhaul extract_digits for 2023.1.2
I think the problem is that the last digit needs to be parsed from the
end, e.g., "abcone234oneight"'s last digit is 8, not 1. When parsing in
order front to back, the 'e' would be consumed and interpreted as a 1,
but apparently the right answer needs to be parsed from the end. In any
case, this meant rewriting the code to separate first and last digit.